Added detail view tests. Refactored test files and split tests up for readability.
