+++ ballots/test_ballotadmin.py
@@ -1,358 +1,12 @@
 import datetime
 from django.contrib.auth.models import User
-from django.test import Client, TestCase, RequestFactory
+from django.test import Client, TestCase
 from django.urls import reverse
 from django.utils import timezone
 
 from .models import Ballot, Question, Choice, VoteRecord
 from users.models import Profile
 from .forms import AddBallotForm, BallotQuestionFormset, QuestionChoiceFormset
-
-class BallotModelTests(TestCase):
-    def setUp(self):
-        self.user = User.objects.create(first_name='John', last_name='Smith', username='testuser',
-                                        password='password123', email='JohnSmith@gmail.com')
-        self.user.save
-        profile = self.user.profile
-        profile.ssn = "555-55-5555"
-        profile.district = "BaltimoreCounty"
-        profile.middle_name = "Jack"
-        self.ballot = Ballot.objects.create(ballot_title="Test", district="BaltimoreCounty", pub_date=timezone.now())
-
-    def test_was_published_recently_with_future_ballot(self):
-        """
-        was_published_recently() returns False for ballots whose pub_date
-        is in the future.
-        """
-        time = timezone.now() + datetime.timedelta(days=30)
-        future_ballot = Ballot(pub_date=time)
-        self.assertIs(future_ballot.was_published_recently(), False)
-
-    def test_was_published_recently_with_old_ballot(self):
-        """
-        was_published_recently() returns False for ballots whose pub_date
-        is older than 1 day.
-        """
-        time = timezone.now() - datetime.timedelta(days=1, seconds=1)
-        old_ballot = Ballot(pub_date=time)
-        self.assertIs(old_ballot.was_published_recently(), False)
-
-    def test_was_published_recently_with_recent_ballot(self):
-        """
-        was_published_recently() returns True for ballots whose pub_date
-        is within the last day.
-        """
-        time = timezone.now() - datetime.timedelta(hours=23, minutes=59, seconds=59)
-        recent_ballot = Ballot(pub_date=time)
-        self.assertIs(recent_ballot.was_published_recently(), True)
-
-    def test_ballot_model_str(self):
-        """
-        Making sure the return str for ballot title is the same str given
-        """
-        ballot = Ballot.objects.create(
-            ballot_title="Test Case Ballot",
-            ballot_description="Making sure the return str is the same",
-            pub_date=timezone.now(),
-            due_date=timezone.now() + datetime.timedelta(days=30)
-        )
-        self.assertEqual(str(ballot), "Test Case Ballot")
-
-    """
-    makes sure the due date of the ballot is after the publish.
-    Don't want a ballot to be past due as soon as it is created.
-    Fixed by clean function in models.py
-    """
-
-
-class QuestionModelTests(TestCase):
-    def test_question_model_str(self):
-        """
-        Making sure the return str for question text is the same str given
-        """
-        ballot = Ballot.objects.create(
-            ballot_title="Test Case Ballot",
-            ballot_description="Making sure the question return str is the same",
-            district="All",
-            pub_date=timezone.now(),
-            due_date=timezone.now() + datetime.timedelta(days=30)
-        )
-        question = Question.objects.create(
-            ballot=ballot,
-            question_text="Make sure the return str is same"
-        )
-        self.assertEqual(str(question), "Make sure the return str is same")
-
-
-class ChoiceModelTests(TestCase):
-    def setUp(self):
-        self.user = User.objects.create(first_name='John', last_name='Smith', username='testuser',
-                                        password='password123', email='JohnSmith@gmail.com')
-        self.user.save
-        profile = self.user.profile
-        profile.ssn = "555-55-5555"
-        profile.district = "BaltimoreCounty"
-        profile.middle_name = "Jack"
-        self.ballot = Ballot.objects.create(ballot_title="Test", district="BaltimoreCounty", pub_date=timezone.now())
-
-    def test_choice_model_str(self):
-        """
-        Making sure the return str for choice text is the same str given
-        """
-        ballot = Ballot.objects.create(
-            ballot_title="Test Case Ballot",
-            ballot_description="Making sure the question return str is the same",
-            district="All",
-            pub_date=timezone.now(),
-            due_date=timezone.now() + datetime.timedelta(days=30)
-        )
-        question = Question.objects.create(
-            ballot=ballot,
-            question_text="Make sure the return str is same"
-        )
-        choice1 = Choice.objects.create(
-            question=question,
-            choice_text="True",
-        )
-        choice2 = Choice.objects.create(
-            question=question,
-            choice_text="False",
-        )
-        self.assertEqual(str(choice1), "True")
-        self.assertEqual(str(choice2), "False")
-
-    def test_choice_vote_count(self):
-        """
-        Making sure the read only votes field can still increment when someone votes
-        """
-        ballot = Ballot.objects.create(
-            ballot_title="Test Case Ballot",
-            ballot_description="Making sure the question return str is the same",
-            district="All",
-            pub_date=timezone.now(),
-            due_date=timezone.now() + datetime.timedelta(days=30)
-        )
-        question = Question.objects.create(
-            ballot=ballot,
-            question_text="Make sure the return str is same"
-        )
-        choice = Choice.objects.create(
-            question=question,
-            choice_text="True",
-            votes=0
-        )
-        self.assertEqual(choice.votes, 0)
-        choice.votes += 1
-        self.assertEqual(choice.votes, 1)
-
-class IndexTests(TestCase):
-    def setUp(self):
-        self.user = User.objects.create(first_name='John', last_name='Smith', username='testuser',
-                                        password='password123', email='JohnSmith@gmail.com')
-        self.user.save
-        profile = self.user.profile
-        profile.ssn = "555-55-5555"
-        profile.district = "BaltimoreCounty"
-        profile.middle_name = "Jack"
-        profile.save()
-        self.ballot = Ballot.objects.create(ballot_title="Test", district="BaltimoreCounty", pub_date=timezone.now())
-        self.ballot.save()
-
-    """index url tests"""
-    def test_index_url_exists(self):
-        """
-        if url exists shouldn't 404, status code should be 302 due to redirect
-        """
-        response = self.client.get('/')
-        self.assertNotEqual(response.status_code, 404)
-        self.assertEqual(response.status_code, 302)
-
-    def test_index_url_reverse(self):
-        """
-        reverse match should exist
-        if url exists shouldn't 404, status code should be 302 due to redirect
-        """
-        response = self.client.get(reverse('ballots:index'))
-        self.assertNotEqual(response.status_code, 404)
-        self.assertEqual(response.status_code, 302)
-    def test_index_anon_redirect(self):
-        """
-        if no user is logged in redirects to login page
-        """
-        response = self.client.get('/')
-        self.assertRedirects(response, '/users/login/')
-    def test_index_login(self):
-        """if user is logged in displays ballot index page"""
-        self.client.force_login(self.user)
-        response = self.client.get('/')
-        self.assertEqual(response.status_code, 200)
-    """end url tests"""
-
-    """index queryset tests"""
-    def test_district_list(self):
-        """
-        ballot_list should only contain published ballots with districts matching the user's district
-        """
-        yesterday = timezone.now() - datetime.timedelta(days=1)
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('ballots:index'))
-
-        """list size should be one"""
-        self.assertEqual(response.context['ballot_list'].count(), 1)
-
-        wrong_ballot = Ballot(ballot_title="Wrong", district="MontgomeryCounty", pub_date=yesterday)
-        wrong_ballot.save()
-        response = self.client.get(reverse('ballots:index'))
-        """list size should not change"""
-        self.assertEqual(response.context['ballot_list'].count(), 1)
-
-        right_ballot = Ballot(ballot_title="Wrong", district=self.user.profile.district, pub_date=yesterday)
-        right_ballot.save()
-        self.ballot.save()
-        response = self.client.get(reverse('ballots:index'))
-        """list size should be two"""
-        self.assertEqual(response.context['ballot_list'].count(), 2)
-        for ballot in response.context['ballot_list']:
-            """all ballot districts should match user district"""
-            self.assertEqual(response.context['user'].profile.district, ballot.district)
-
-    def test_date_list(self):
-        """
-        ballot_list should only contain published ballots with districts matching the user's district
-        """
-        tomorrow = timezone.now() + datetime.timedelta(days=1)
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('ballots:index'))
-
-        """list size should be one"""
-        self.assertEqual(response.context['ballot_list'].count(), 1)
-
-        wrong_ballot = Ballot(ballot_title="Wrong", district=self.user.profile.district, pub_date=tomorrow)
-        wrong_ballot.save()
-        response = self.client.get(reverse('ballots:index'))
-        """list size should not change"""
-        self.assertEqual(response.context['ballot_list'].count(), 1)
-
-        right_ballot = Ballot(ballot_title="Wrong", district=self.user.profile.district, pub_date=timezone.now())
-        right_ballot.save()
-        self.ballot.save()
-        response = self.client.get(reverse('ballots:index'))
-        """list size should be two"""
-        self.assertEqual(response.context['ballot_list'].count(), 2)
-        for ballot in response.context['ballot_list']:
-            """all ballot districts should match user district"""
-            self.assertLessEqual(ballot.pub_date, timezone.now())
-
-    def test_finished_list(self):
-        """
-        ballot_list should only contain ballots user has not voted on
-        """
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('ballots:index'))
-
-        """list size should be one"""
-        self.assertEqual(response.context['ballot_list'].count(), 1)
-        """finished list size should be zero"""
-        self.assertEqual(response.context['finished_ballots'].count(), 0)
-
-        finished_ballot = Ballot(ballot_title="Wrong", district=self.user.profile.district, pub_date=timezone.now())
-        finished_ballot.save()
-        response = self.client.get(reverse('ballots:index'))
-        """list size should not change"""
-        self.assertEqual(response.context['ballot_list'].count(), 2)
-        """finished list size should not change"""
-        self.assertEqual(response.context['finished_ballots'].count(), 0)
-
-        vote_record = VoteRecord(voter_signature=self.user.profile.sign, assoc_ballot=finished_ballot)
-        vote_record.save()
-        response = self.client.get(reverse('ballots:index'))
-        """list size should be one"""
-        self.assertEqual(response.context['ballot_list'].count(), 1)
-        """finished list size should be one"""
-        self.assertEqual(response.context['finished_ballots'].count(), 1)
-        for ballot in response.context['ballot_list']:
-            """only test ballot should be in this list"""
-            self.assertEqual(ballot.ballot_title, "Test")
-        for ballot in response.context['finished_ballots']:
-            """only wrong ballot should be in this list"""
-            self.assertEqual(ballot.ballot_title, "Wrong")
-    """end queryset tests"""
-
-class BallotFormTests(TestCase):
-    """
-    test ballot form labels
-    """
-    def test_ballot_form_title(self):
-        form = AddBallotForm()
-        self.assertTrue(form.fields['ballot_title'].label is None or form.fields['ballot_title'].
-                        label == 'Ballot title')
-
-    def test_ballot_form_description(self):
-        form = AddBallotForm()
-        self.assertTrue(form.fields['ballot_description'].label is None or form.fields['ballot_description'].
-                        label == 'Ballot description')
-
-    def test_ballot_form_district(self):
-        form = AddBallotForm()
-        self.assertTrue(form.fields['ballot_description'].label is None or form.fields['district'].
-                        label == 'District')
-
-    def test_ballot_form_pub_date(self):
-        form = AddBallotForm()
-        self.assertTrue(form.fields['pub_date'].label is None or form.fields['pub_date'].
-                        label == 'Date published')
-
-    def test_ballot_form_due_date(self):
-        form = AddBallotForm()
-        self.assertTrue(form.fields['due_date'].label is None or form.fields['due_date'].
-                        label == 'Due date')
-
-    """
-    end label tests
-    """
-
-    """
-    form validation tests
-    """
-    def test_ballot_form_pub_past(self):
-        """
-        forms cannot be published in the past
-        """
-        date = timezone.now() - datetime.timedelta(days=1)
-        form = AddBallotForm(data={'ballot_title': "Test", 'ballot_description': "test", 'pub_date': date,
-                                   'due_date': timezone.now() + datetime.timedelta(days=30)})
-        self.assertFalse(form.is_valid())
-
-    def test_ballot_form_pub_now(self):
-        """
-        forms cannot be published now
-        """
-        date = timezone.now()
-        form = AddBallotForm(data={'ballot_title': "Test", 'ballot_description': "test", 'pub_date': date,
-                                   'due_date': timezone.now() + datetime.timedelta(days=30)})
-        self.assertFalse(form.is_valid())
-
-    def test_ballot_form_pub_future(self):
-        """
-        forms can be published in the future
-        """
-        date = timezone.now() + datetime.timedelta(days=1)
-        form = AddBallotForm(data={'ballot_title': "Test", 'ballot_description': "test", 'pub_date':date,
-                                   'due_date':timezone.now() + datetime.timedelta(days=30)})
-        self.assertTrue(form.is_valid())
-
-    def test_ballot_form_pub_after_due(self):
-        """
-        form publish date must be before form due date
-        """
-        date = timezone.now() + datetime.timedelta(days=31)
-        form = AddBallotForm(data={'ballot_title': "Test", 'ballot_description': "test", 'pub_date':date,
-                                   'due_date': timezone.now() + datetime.timedelta(days=30)})
-        self.assertFalse(form.is_valid())
-
-    """
-    end validation tests
-    """
 
 class BallotAdminTests(TestCase):
     def setUp(self):
